Move tests related to Lock and Event into their own module
